# hexabid-api-contract/src/main/resources/openapi/auction-api.yaml
openapi: 3.0.3
info:
  title: Auctions API
  version: 1.0.0
servers:
  - url: http://localhost:18080/hexabid
...
paths:
  /api/products:
    get:
      operationId: browseProductTypes
      tags:
        - Products
      summary: Browse all product types in catalog
      description: |
        Returns paginated list of all product types available in the system.
        This is the main entry point for displaying the product catalog.
        
        **Archetypowy kontekst:** Product (M02 - Katalog Produktów)
        Produkt to definicja "co to jest" - niezależny kontekst od Inventory.
        ProductType reprezentuje szablon, według którego tworzone są instancje.
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - in: query
          name: query
          description: Filter by product name (partial match, case-insensitive)
          schema:
            type: string
        - in: query
          name: trackingStrategy
          description: Filter by product tracking strategy
          schema:
            $ref: '#/components/schemas/ProductTrackingStrategy'
        - in: query
          name: limit
          description: Maximum number of items to return
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - in: query
          name: after
          description: Cursor for pagination - returns items after this cursor
          schema:
            type: string
      responses:
        '200':
          description: Product type list (paginated)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductTypeListResponse'
        '400':
          description: Invalid request parameters
    post:
      operationId: createProductType
      tags:
        - Products
      summary: Define a new product type
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductTypeRequest'
      responses:
        '201':
          description: Product type created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductTypeResponse'
        '400':
          description: Invalid request
  /api/products/{productId}:
    get:
      operationId: getProductType
      tags:
        - Products
      summary: Get product type by ID
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - in: path
          name: productId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product type details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductTypeResponse'
        '404':
          description: Product type not found
...
components:
  parameters:
    ApiVersionHeader:
      in: header
      name: X-API-Version
      required: false
      schema:
        type: string
        default: "1"
      description: API version negotiated via HTTP header.
  schemas:
...
    ProductTrackingStrategy:
      type: string
      enum:
        - UNIQUE
        - INDIVIDUALLY_TRACKED
        - BATCH_TRACKED
        - INDIVIDUALLY_AND_BATCH_TRACKED
        - IDENTICAL
    CreateProductTypeRequest:
      type: object
      required:
        - name
        - description
        - trackingStrategy
        - preferredUnit
      properties:
        name:
          type: string
          example: "Jan Kowalski's Car"
        description:
          type: string
          example: "Seat Leon 1999 TDI"
        trackingStrategy:
          $ref: '#/components/schemas/ProductTrackingStrategy'
        preferredUnit:
          type: string
          example: "pcs"
    ProductTypeResponse:
      type: object
      required:
        - productId
        - name
        - trackingStrategy
        - preferredUnit
      properties:
        productId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        trackingStrategy:
          $ref: '#/components/schemas/ProductTrackingStrategy'
        preferredUnit:
          type: string
...
    ProductTypeListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          description: List of product types
          items:
            $ref: '#/components/schemas/ProductTypeResponse'
        nextCursor:
          type: string
          description: Cursor for pagination - pass this in next request to get next page
          nullable: true
...